requirements.txt
streamlit
tesseractOCR
pytesseract
poppler
Pillow
pdf2image
configparser
pathlib
python-dotenv
google-generativeai
pandas